requirements.txt
chromadb
streamlit
google-generativeai
python-dotenv
PyPDF2